add figure 2.6

+++ harness/src/main/scala/com/robertjneal/rl/actionselector/ActionSelectors.scala
@@ -30,7 +30,7 @@ def softMax(actionPreferences: Map[Action, Preference]): Action = {
   pickWithProbabilty[Action](Probability.random, actionProbabilities)
 }
 
-def upperConfidenceBound(c: Int, state: State)(step: Step, actionSteps: Map[State, Map[Action, Step]])(actionRewards: Map[Action, Reward]): Action = {
+def upperConfidenceBound(c: Double, state: State)(step: Step, actionSteps: Map[State, Map[Action, Step]])(actionRewards: Map[Action, Reward]): Action = {
   def valueTransformer(action: Action, reward: Reward): Double = {
     val actionStep = actionSteps.getOrElse(state, Map(action -> Step(1))).getOrElse(action, Step(1))
     reward.toDouble + c * Math.sqrt(Math.log(step.toInt)/actionStep.toInt.toDouble)
